Created a helper for send_borrowing_notifications
